[AN] refactor: SearchedHearit 모델 통합 및 검색 결과 아이템 컴포넌트 공통화
* [AN] refactor: SearchedCategoryHearit 모델을 삭제하고 SearchedHearit으로 통합
* [AN] refactor: SearchedHearit에 createdAt, category 필드 추가
* [AN] refactor: 기존 SearchResultItem과 SearchedHearitItem을 SearchedHearitItem으로 통합 및 위치 변경 (search/component)
* [AN] refactor: HearitTitle, HearitMetaRow를 별도 파일에서 SearchedHearitItem 내 private 컴포넌트로 변경
* [AN] refactor: CategoryScreen 및 SearchResultScreen에서 통합된 SearchedHearitItem을 사용하도록 수정
* [AN] refactor: HearitRepository 및 Mapper에서 통합된 SearchedHearit 모델을 사용하도록 변경
